Add form layout options
Couldn't get this to actually work but I'm pretty sure this is how it is
meant to work.

I suspect I need to setup the toolbar for the InnerBlocks / form block.
That should let me see the layout option in the toolbar.

It will probably also clean up the weirdness where the pinned toolbar
keeps focus of the previous selected block even though props.focus has
updated to the form block (as per inspector controls showing)